cleaning refactoring
